+++ tests/test_jump.py
@@ -32,6 +32,17 @@ def setup_cube():
     return _cube
 
 
+def test_find_simple_circle():
+    plane = np.zeros(shape=(5, 5), dtype=np.uint8)
+    plane[2, 2] = DQFLAGS['JUMP_DET']
+    plane[3, 2] = DQFLAGS['JUMP_DET']
+    plane[1, 2] = DQFLAGS['JUMP_DET']
+    plane[2, 3] = DQFLAGS['JUMP_DET']
+    plane[2, 1] = DQFLAGS['JUMP_DET']
+    circle = find_circles(plane, DQFLAGS['JUMP_DET'], 1)
+    assert circle[0][1] == pytest.approx(1.0, 1e-3)
+
+
 def test_find_simple_ellipse():
     plane = np.zeros(shape=(5, 5), dtype=np.uint8)
     plane[2, 2] = DQFLAGS['JUMP_DET']